feat(dogecoin): add difficulty adjustment algorithm for blocks 0-5,000 (#8)
This PR adds methods required for implementing Dogecoin's difficulty
adjustment algorithm (DAA). In particular it adds the method
`from_next_work_required_dogecoin()` which computes the target value for
the next difficulty adjustment period, based on the `timespan` between
start and end of period and the `last` target value.

Note: this PR only adds the required logic to compute targets in block
period 0-5,000 (see
https://github.com/dogecoin/dogecoin/blob/51cbc1fd5d0d045dda2ad84f53572bbf524c6a8e/src/dogecoin.cpp#L41
for complete algorithm). A following PR will add the logic for further
periods and in particular for the Digishield algorithm used in later
periods.